Refactor database cloning logic into CosmosService and add configuration management
