+++ CosmosSchemaCloner.csproj
@@ -12,4 +12,12 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Content Include=".github\workflows\release.yml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="publish\" />
+  </ItemGroup>
+
 </Project>
